Black Jack+ Time Series+ Best Regression FSS
Black Jack project
We would like to see our expected winnings (losses) from playing the
casino card game, Blackjack (https://en.wikipedia.org/wiki/Blackjack) .
We want to see a basic strategy and also a card counting strategy. Each
of the two strategies' efficacies will be evaluated with a simulation
and visualized with a histogram plot (i.e., there should be two
figures­­one for each strategy).

Time Series project
The Gold Plan
So we heard that an inverted yield curve is a signal for recession. And
then we heard that the best thing to hold during a recession is gold!
Let's do some analysis to see if that might offer a good trading
strategy.

Best Regression FSS
In this project assignment, we will use regression analysis in order to
do feature set selection. This technique is generally an
exponential­time algorithm so is intractable for anything but very
small feature dimensionality (p < 30). We will use a brute­force
approach to find the most appealing feature set